working with permission wise page view

+++ app/Http/Controllers/Backend/RolesController.php
@@ -5,11 +5,22 @@ namespace App\Http\Controllers\Backend;
 use App\Http\Controllers\Controller;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
 
 class RolesController extends Controller
 {
+
+    public $user;
+
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::guard('admin')->user();
+            return $next($request);
+        });
+    } //end of the constructor method
     /**
      * Display a listing of the resource.
      *
@@ -17,8 +28,12 @@ class RolesController extends Controller
      */
     public function index()
     {
-        $roles = Role::all();
+        if (is_null($this->user) || !$this->user->can('role.view')) {
+            abort(403, 'Sorry !! You are Unauthorized to view role list !');
 
+        }
+
+        $roles = Role::all();
         return view('backend.pages.roles.index', compact('roles'));
     } //end of the index method
 
@@ -29,6 +44,10 @@ class RolesController extends Controller
      */
     public function create()
     {
+        if (is_null($this->user) || !$this->user->can('role.create')) {
+            abort(403, 'Sorry !! You are Unauthorized to create any role !');
+        }
+
         $all_permissions = Permission::all();
 
         $permission_groups = User::getpermissionsGroups();
@@ -44,6 +63,9 @@ class RolesController extends Controller
      */
     public function store(Request $request)
     {
+        if (is_null($this->user) || !$this->user->can('role.create')) {
+            abort(403, 'Sorry !! You are Unauthorized to store any role !');
+        }
 
         //data validation
         $request->validate([
@@ -53,7 +75,7 @@ class RolesController extends Controller
         ]);
 
         //data processing
-        $role = Role::create(['name' => $request->name]);
+        $role = Role::create(['name' => $request->name, 'guard_name' => 'admin']);
 
         // $role = DB::table('roles')->where('name', $request->name)->first();
         $permissions = $request->input('permissions');
@@ -85,7 +107,11 @@ class RolesController extends Controller
      */
     public function edit($id)
     {
-        $role = Role::findById($id);
+        if (is_null($this->user) || !$this->user->can('role.edit')) {
+            abort(403, 'Sorry !! You are Unauthorized to edit any role !');
+        }
+
+        $role = Role::findById($id, 'admin');
 
         $all_permissions = Permission::all();
 
@@ -104,6 +130,10 @@ class RolesController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (is_null($this->user) || !$this->user->can('role.edit')) {
+            abort(403, 'Sorry !! You are Unauthorized to update any role !');
+        }
+
         //data validation
         $request->validate([
             'name' => 'required|max:100|unique:roles,name,' . $id,
@@ -111,7 +141,7 @@ class RolesController extends Controller
             'name.required' => 'Please give a role name',
         ]);
 
-        $role = Role::findById($id);
+        $role = Role::findById($id, 'admin');
 
         $permissions = $request->input('permissions');
 
@@ -134,8 +164,11 @@ class RolesController extends Controller
      */
     public function destroy($id)
     {
+        if (is_null($this->user) || !$this->user->can('role.delete')) {
+            abort(403, 'Sorry !! You are Unauthorized to delete any role !');
+        }
 
-        $role = Role::findById($id);
+        $role = Role::findById($id, 'admin');
 
         if (!is_null($role)) {
             $role->delete();
